knotserver: /init and /user endpoints
When the knot is initialized, the h.init channel is closed signalling
the Jetstream watcher to begin, with the owner DID in filter. There's no
null filter so starting it without any DIDs will give us events for all
DIDs. /init also checks if h.knotInitialized is true, and will 409
Conflict if /init is called again.

PUT /user is similar, but will only trigger a call to UpdateDids.