jestream: allow caller to optionally waitForDid
This is handy because the knotserver only subscribes to specific dids,
and we don't want jestream to run without (which defaults to all dids).
The oppsite is true for the appview which wants to see all dids.

Also adds some nil checks to ensure it doesn't break when event is not a
Commit.